// shopping-cart/src/app/models/product.ts
//This is product model
export class Product {
    id: number;
    name: string;
    description: string;
    price: number;
    imageUrl: string;

    // constructor(id, name, description = '', price = 0, imageUrl = 'https://scx2.b-cdn.net/gfx/news/hires/2018/1-rubikscube.jpg') {
    //     this.id = id;
    //     this.name = name;
    //     this.description = description;
    //     this.price = price;
    //     this.imageUrl = imageUrl;
    // }
}